stm32/hrtim: split off BridgeConverter and ResonantConverter converter implementations to separate modules
